Organize for new structure
